set logging
