+++ src/digrec/iopack/Reader.java
@@ -1,10 +1,11 @@
-package iopack;
+package digrec.iopack;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.logging.Level;
 
 public class Reader {
 	String name;
@@ -14,6 +15,7 @@ public class Reader {
 		name = path;
 		Path pth = Paths.get(path);
 				if(Files.exists(pth)) {
+					Test.LOG.log(Level.FINE, "Файл найден");
 					this.storage = Files.readAllBytes(pth);
 				}
 				else {
