+++ src/logging.properties
@@ -1,0 +1,9 @@
+#handlers= java.util.logging.FileHandler
+
+handlers= java.util.logging.ConsoleHandler
+.level=ALL
+java.util.logging.ConsoleHandler.level = ALL
+
+java.util.logging.FileHandler.pattern = log
+java.util.logging.FileHandler.limit = 1000000
+java.util.logging.FileHandler.count = 7
